requirements.txt
python-dotenv==1.0.1
groq>=0.11.0
httpx==0.27.2